Expose Resource handle variant and update usage
Added GetHandleVariant() to ResourceVariant for direct access to the handle variant. Updated VariadicTypedNode and DescriptorResourceGathererNode to use the new method for extracting resource handles. Fixed type conversion in RenderGraph cleanup completion event.

+++ VIXEN/RenderGraph/src/Core/RenderGraph.cpp
@@ -1244,7 +1244,7 @@ void RenderGraph::HandleCleanupRequest(const EventTypes::CleanupRequestedMessage
 
     // Publish completion event
     if (messageBus) {
-        auto completionMsg = std::make_unique<EventTypes::CleanupCompletedMessage>(0, cleanedCount);
+        auto completionMsg = std::make_unique<EventTypes::CleanupCompletedMessage>(0, static_cast<uint32_t>(cleanedCount));
         messageBus->Publish(std::move(completionMsg));
     }
 }
